+++ src/main/java/com/silkaitis/pricing/PriceCalculator.java
@@ -3,6 +3,7 @@ package com.silkaitis.pricing;
 import com.silkaitis.exception.LogicException;
 import com.silkaitis.model.Product;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface PriceCalculator {
@@ -12,5 +13,5 @@ public interface PriceCalculator {
      * @param products - List of products.
      * @return - Total price.
      */
-    Double calculateTotalPrice(List<Product> products) throws LogicException;
+    BigDecimal calculateTotalPrice(List<Product> products) throws LogicException;
 }
